Remove TODO: because lib/pq now does this
The only thing is the connection string must start with postgres://
not postgresql:// that won't work.

This was recently merged in to the lib/pq library, so we can already
use 12-factor style DB connection URLs right now.

It might be worth to send a patch upstream to allow postgresql://
as well as postgres://