Move migrate to command, add createsuperuser command
For now, the hash algorithm used is pbkdf2-sha256 though the ability
to change this to bcrypt will be nice to add later.

For the time being, the createsuperuser command will not ask for user
details and always generate a user called "admin" with password "password".
This is temporary and will change.